requirements.txt
streamlit>=1.25.0
pillow
numpy
onnxruntime
